refactor(ai): split 681-LOC rag_service into retriever/attribution/support
Recipe R2: extract RagRetriever (rewrite→expand→embed→search→RRF→resolve),
RagAttribution (source titles/debug notes), RagSupport (messages, error
mapping, chat-memory prep, debug formatting), and RagResult into their own
files. RagService stays the orchestrator at 294 LOC. Public API unchanged —
RagService.rrfFuse kept as a static delegate; constructor + ServiceLocator
getter signatures preserved. All RAG/chat/quiz tests pass.
